requirements.txt
# packages in environment at /anaconda3/envs/mauda:
#
# Name                    Version                   Build  Channel
_libgcc_mutex             0.1                        main    defaults
_openmp_mutex             4.5                       1_gnu    defaults
addict                    2.4.0            py37h89c1867_0    conda-forge
blas                      1.0                         mkl    defaults
bzip2                     1.0.8                h7f98852_4    conda-forge
ca-certificates           2021.10.8            ha878542_0    conda-forge
cairo                     1.16.0            h18b612c_1001    conda-forge
certifi                   2021.10.8        py37h89c1867_0    conda-forge
click                     8.1.3                    pypi_0    pypi
cudatoolkit               10.1.243             h6bb024c_0    defaults
cycler                    0.10.0                   pypi_0    pypi
deprecated                1.2.13                   pypi_0    pypi
ffmpeg                    4.0                  hcdf2ecd_0    defaults
fontconfig                2.13.1            hba837de_1005    conda-forge
freeglut                  3.2.1                h9c3ff4c_2    conda-forge
freetype                  2.10.4               h5ab3b9f_0    defaults
future                    0.18.2                   pypi_0    pypi
glib                      2.69.1               h5202010_0    defaults
graphite2                 1.3.13            h58526e2_1001    conda-forge
harfbuzz                  1.8.8                hffaf4a1_0    defaults
hdf5                      1.10.2               hc401514_3    conda-forge
humanize                  4.2.3                    pypi_0    pypi
icu                       58.2              hf484d3e_1000    conda-forge
importlib-metadata        4.12.0                   pypi_0    pypi
intel-openmp              2021.3.0          h06a4308_3350    defaults
jasper                    2.0.14               hd8c5072_2    defaults
jpeg                      9d                   h7f8727e_0    defaults
kiwisolver                1.3.2                    pypi_0    pypi
lcms2                     2.12                 h3be6417_0    defaults
ld_impl_linux-64          2.35.1               h7274673_9    defaults
libffi                    3.3                  he6710b0_2    defaults
libgcc-ng                 9.3.0               h5101ec6_17    defaults
libgfortran               3.0.0                         1    conda-forge
libglu                    9.0.0             he1b5a44_1001    conda-forge
libgomp                   9.3.0               h5101ec6_17    defaults
libopencv                 3.4.2                hb342d67_1    defaults
libopus                   1.3.1                h7f98852_1    conda-forge
libpng                    1.6.37               hbc83047_0    defaults
libstdcxx-ng              9.3.0               hd4cf53a_17    defaults
libtiff                   4.2.0                h85742a9_0    defaults
libuuid                   2.32.1            h7f98852_1000    conda-forge
libvpx                    1.7.0                h439df22_0    defaults
libwebp-base              1.2.0                h27cfd23_0    defaults
libxcb                    1.13              h7f98852_1003    conda-forge
libxml2                   2.9.12               h03d6c58_0    defaults
lz4-c                     1.9.3                h295c915_1    defaults
matplotlib                3.4.3                    pypi_0    pypi
mkl                       2021.3.0           h06a4308_520    defaults
mkl-service               2.4.0            py37h7f8727e_0    defaults
mkl_fft                   1.3.0            py37h42c9631_2    defaults
mkl_random                1.2.2            py37h51133e4_0    defaults
mmcv-full                 1.2.2                    pypi_0    pypi
ncurses                   6.2                  he6710b0_1    defaults
nibabel                   3.2.1                    pypi_0    pypi
ninja                     1.10.2               hff7bd54_1    defaults
numpy                     1.21.2           py37h20f2e39_0    defaults
numpy-base                1.21.2           py37h79a1101_0    defaults
olefile                   0.46                     py37_0    defaults
opencv                    3.4.2            py37h6fd60c2_1    defaults
opencv-python             4.5.3.56                 pypi_0    pypi
openjpeg                  2.4.0                h3ad879b_0    defaults
openssl                   1.1.1l               h7f8727e_0    defaults
packaging                 21.0                     pypi_0    pypi
pcre                      8.45                 h9c3ff4c_0    conda-forge
pillow                    8.3.1            py37h2c7a002_0    defaults
pip                       21.0.1           py37h06a4308_0    defaults
pixman                    0.38.0            h516909a_1003    conda-forge
pthread-stubs             0.4               h36c2ea0_1001    conda-forge
py-opencv                 3.4.2            py37hb342d67_1    defaults
pyparsing                 2.4.7                    pypi_0    pypi
python                    3.7.11               h12debd9_0    defaults
python-dateutil           2.8.2                    pypi_0    pypi
python_abi                3.7                     2_cp37m    conda-forge
pytorch                   1.6.0           py3.7_cuda10.1.243_cudnn7.6.3_0    pytorch
pyyaml                    5.4.1            py37h5e8e339_0    conda-forge
readline                  8.1                  h27cfd23_0    defaults
scipy                     1.7.1                    pypi_0    pypi
setuptools                58.0.4           py37h06a4308_0    defaults
simpleitk                 2.1.1                    pypi_0    pypi
six                       1.16.0             pyhd3eb1b0_0    defaults
sqlite                    3.36.0               hc218d9a_0    defaults
terminaltables            3.1.0                    pypi_0    pypi
timm                      0.4.12                   pypi_0    pypi
tk                        8.6.11               h1ccaba5_0    defaults
torchio                   0.18.82                  pypi_0    pypi
torchvision               0.7.0                py37_cu101    pytorch
tqdm                      4.64.0                   pypi_0    pypi
typing-extensions         4.3.0                    pypi_0    pypi
wheel                     0.37.0             pyhd3eb1b0_1    defaults
wrapt                     1.14.1                   pypi_0    pypi
xorg-fixesproto           5.0               h7f98852_1002    conda-forge
xorg-inputproto           2.3.2             h7f98852_1002    conda-forge
xorg-kbproto              1.0.7             h7f98852_1002    conda-forge
xorg-libice               1.0.10               h7f98852_0    conda-forge
xorg-libsm                1.2.3             hd9c2040_1000    conda-forge
xorg-libx11               1.7.2                h7f98852_0    conda-forge
xorg-libxau               1.0.9                h7f98852_0    conda-forge
xorg-libxdmcp             1.1.3                h7f98852_0    conda-forge
xorg-libxext              1.3.4                h7f98852_1    conda-forge
xorg-libxfixes            5.0.3             h7f98852_1004    conda-forge
xorg-libxi                1.7.10               h7f98852_0    conda-forge
xorg-libxrender           0.9.10            h7f98852_1003    conda-forge
xorg-renderproto          0.11.1            h7f98852_1002    conda-forge
xorg-xextproto            7.3.0             h7f98852_1002    conda-forge
xorg-xproto               7.0.31            h7f98852_1007    conda-forge
xz                        5.2.5                h7b6447c_0    defaults
yaml                      0.2.5                h516909a_0    conda-forge
yapf                      0.31.0             pyhd8ed1ab_0    conda-forge
zipp                      3.8.0                    pypi_0    pypi
zlib                      1.2.11               h7b6447c_3    defaults
zstd                      1.4.9                haebb681_0    defaults